+++ src/gateways/InclusiveGateway.js
@@ -1,13 +1,12 @@
 import Activity from '../activity/Activity';
-import {ActivityError} from '../error/Errors';
-import {cloneContent, cloneMessage} from '../messageHelper';
+import {cloneContent} from '../messageHelper';
 
 export default function InclusiveGateway(activityDef, context) {
   return Activity(InclusiveGatewayBehaviour, activityDef, context);
 }
 
 export function InclusiveGatewayBehaviour(activity) {
-  const {id, type, broker, logger, outbound: outboundSequenceFlows = []} = activity;
+  const {id, type, broker} = activity;
 
   const source = {
     id,
@@ -17,31 +16,7 @@ export function InclusiveGatewayBehaviour(activity) {
 
   return source;
 
-  function execute(executeMessage) {
-    const content = cloneContent(executeMessage.content);
-    if (!outboundSequenceFlows.length) return complete();
-
-    return activity.evaluateOutbound(cloneMessage(executeMessage), false, complete);
-
-    function complete(err, outbound) {
-      if (err) {
-        const error = new ActivityError(err.message, executeMessage, err);
-        return broker.publish('execution', 'execute.error', {...content, error});
-      }
-      if (!outbound) return broker.publish('execution', 'execute.completed', content);
-
-      const taken = outbound.find(({action}) => action === 'take');
-      if (!taken) {
-        const error = new ActivityError(`<${id}> no conditional flow taken`, executeMessage);
-        logger.error(`<${id}>`, err);
-        return broker.publish('execution', 'execute.error', {...content, error});
-      }
-
-      if (taken.isDefault) {
-        logger.debug(`<${id}> take default flow <${taken.id}>`);
-      }
-
-      return broker.publish('execution', 'execute.completed', {...content, outbound});
-    }
+  function execute({content}) {
+    broker.publish('execution', 'execute.completed', cloneContent(content));
   }
 }